+++ docs/changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- Add tests module for sc-agents-utils
 - Add opportunity to configure sc-machine version from CMakeLists.txt
 - Add scripts to build kb, run sctp-server and run sc-server
 - New flag options for build_kb